requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.1.13
django-rest-framework==0.1.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.7.0
docker==7.1.0
idna==3.10
pymongo==4.10.1
pywin32==308
requests==2.32.3
sqlparse==0.2.4
tzdata==2024.2
urllib3==2.2.3
